requirements.txt
webapp2==3.0.0b1
gunicorn==20.1.0
google-cloud-ndb==1.11.2
jinja2==3.1.2
google-cloud-storage
wtforms==3.0.0